\name{codetemplate-class}
\docType{class}
\alias{codetemplate-class}

\title{Class "codetemplate"}
\description{
A `codetemplate' contains a \code{snippetlist},
which can be merged with `templates'.

Used internally by \code{\link{Umacs}}

}
\section{Objects from the Class}{}
\section{Slots}{
	 \describe{
    \item{\code{snippets}:}{Object of class \code{"snippetlist"} ~~ }
  }
}
\section{Extends}{}
\section{Methods}{
  \describe{
	 }
}
\references{Kerman, Jouni. Umacs: A Universal Markov Chain Sampler. Technical report, Columbia University, New York.}
\author{Jouni Kerman <jouni@kerman.com>}
\note{}
\seealso{
\code{\link{Umacs-package}} for an overview of the Umacs package.

\code{\link{Sampler}} for how to create a sampler function; see the vignette for details.

Package `rv'.

}
\examples{}
\keyword{iteration}
\details{More documentation follows (possibly) in later versions.}
