\name{codetemplatelist-class}
\docType{class}
\alias{codetemplatelist-class}

\title{Class "codetemplatelist"}
\description{
A `codetemplatelist' contains a list of \code{codetemplate}s,
containing snippets that can be merged with `templates' at once.

Used internally by \code{\link{Umacs}}
}
\section{Objects from the Class}{}
\section{Slots}{
	 \describe{
    \item{\code{slist}:}{Object of class \code{"list"} ~~ }
  }
}
\section{Extends}{}
\section{Methods}{
  \describe{
	 }
}
\references{Kerman, Jouni. Umacs: A Universal Markov Chain Sampler. Technical report, Columbia University, New York.}
\author{Jouni Kerman <jouni@kerman.com>}
\note{}
\seealso{
\code{\link{Umacs-package}} for an overview of the Umacs package.

\code{\link{Sampler}} for how to create a sampler function; see the vignette for details.

Package `rv'.

}
\examples{}
\keyword{iteration}
\details{
More documentation follows (possibly) in later versions.

}
