\name{ParameterList-class}
\docType{class}
\alias{ParameterList-class}

\title{Class definition of "ParameterList" (Umacs package)}
\description{This class is used only internally, in Umacs.}
\section{Objects from the Class}{}
\section{Slots}{
	 \describe{
    \item{\code{params}:}{Object of class \code{"list"} ~~ }
    \item{\code{check.ok}:}{Object of class \code{"logical"} ~~ }
    \item{\code{slist}:}{Object of class \code{"list"} ~~ }
  }
}
\section{Extends}{}
\section{Methods}{
  \describe{
	 }
}
\references{Kerman, Jouni. Umacs: A Universal Markov Chain Sampler. Technical report, Columbia University, New York.}
\author{Jouni Kerman <jouni@kerman.com>}
\note{}
\seealso{
\code{\link{Umacs-package}} for an overview of the Umacs package.

\code{\link{Sampler}} for how to create a sampler function; see the vignette for details.

Package `rv'.

}
\examples{}
\keyword{iteration}
\details{
Wrapper for a list of ``\code{Parameter}" objects.

}
