\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,mcts-method}
\title{Show a summary of an mcts object}
\description{Displays the summary of the simulations stored inside an mcts object.}
\section{Methods}{
\describe{

\item{obj = "mcts"}{ Displays the summary of the simulations stored inside an mcts object. }

}}
\keyword{iteration}
