\name{snippet-class}
\docType{class}
\alias{snippet-class}

\title{Class "snippet"}
\description{
A class whose instances accommodate ``code snippets"
created by substitute() or quote() etc.

Used internally by \code{\link{Umacs}}
}
\section{Objects from the Class}{}
\section{Slots}{
	 \describe{
    \item{\code{code}:}{Object of class \code{"list"} ~~ }
    \item{\code{args}:}{Object of class \code{"list"} ~~ }
    \item{\code{name}:}{Object of class \code{"character"} ~~ }
    \item{\code{output}:}{Object of class \code{"numeric"} ~~ }
  }
}
\section{Extends}{}
\section{Methods}{
  \describe{
	 }
}
\references{Kerman, Jouni. Umacs: A Universal Markov Chain Sampler. Technical report, Columbia University, New York.}
\author{Jouni Kerman <jouni@kerman.com>}
\note{}
\seealso{
\code{\link{Umacs-package}} for an overview of the Umacs package.

\code{\link{Sampler}} for how to create a sampler function; see the vignette for details.

Package `rv'.

}
\examples{}
\keyword{iteration}
\details{}
