\name{snippetlist-class}
\docType{class}
\alias{snippetlist-class}

\title{Class "snippetlist"}
\description{
A class whose instance accommodates a list of ``code snippets" of class ``snippet."

Used internally by \code{\link{Umacs}}
}
\section{Objects from the Class}{}
\section{Slots}{
	 \describe{
    \item{\code{snippets}:}{Object of class \code{"list"} ~~ }
  }
}
\section{Extends}{}
\section{Methods}{
  \describe{
	 }
}
\references{Kerman, Jouni. Umacs: A Universal Markov Chain Sampler. Technical report, Columbia University, New York.}
\author{Jouni Kerman <jouni@kerman.com>}
\note{}
\seealso{
\code{\link{Umacs-package}} for an overview of the Umacs package.

\code{\link{Sampler}} for how to create a sampler function; see the vignette for details.

Package `rv'.

}
\examples{}
\keyword{iteration}
\details{}
