\name{snippets-methods}
\docType{methods}
\alias{snippets-methods}
\alias{snippets,snippetlist-method}
\alias{snippets,codetemplate-method}
\alias{snippets,codetemplatelist-method}
\alias{snippets}
\title{Get the `snippets' of an object}
\description{Used internally in \code{\link{Umacs}}.}
\section{Methods}{
\describe{

\item{obj = "snippetlist"}{ ~~describe this method here }

\item{obj = "codetemplate"}{ ~~describe this method here }

\item{obj = "codetemplatelist"}{ ~~describe this method here }
}}
\keyword{iteration}
