\name{templates-methods}
\docType{methods}
\alias{templates-methods}
\alias{templates,codetemplate-method}
\alias{templates,Parameter-method}
\alias{templates,SamplingScheme-method}
\alias{templates,RawCode-method}
\alias{templates,Gibbs-method}
\alias{templates,Data-method}
\alias{templates,DFunction-method}
\alias{templates,LocalObject-method}
\alias{templates,LocalFunction-method}
\alias{templates,ParameterList-method}
\alias{templates,Sampler-method}
\alias{templates,SamplerCheck-method}
\alias{templates,Trace-method}
\alias{templates,Metropolis-method}
\alias{templates,SMetropolis-method}
\alias{templates,PSMetropolis-method}
\alias{templates,PMetropolis-method}
\alias{templates}
\title{Get the `templates' of an object}
\description{Used internally in \code{\link{Umacs}}.}
\section{Methods}{
\describe{

\item{obj = "codetemplate"}{ ~~describe this method here }

\item{obj = "Parameter"}{ ~~describe this method here }

\item{obj = "SamplingScheme"}{ ~~describe this method here }

\item{obj = "RawCode"}{ ~~describe this method here }

\item{obj = "Gibbs"}{ ~~describe this method here }

\item{obj = "Data"}{ ~~describe this method here }

\item{obj = "DFunction"}{ ~~describe this method here }

\item{obj = "LocalObject"}{ ~~describe this method here }

\item{obj = "LocalFunction"}{ ~~describe this method here }

\item{obj = "ParameterList"}{ ~~describe this method here }

\item{obj = "Sampler"}{ ~~describe this method here }

\item{obj = "SamplerCheck"}{ ~~describe this method here }

\item{obj = "Trace"}{ ~~describe this method here }

\item{obj = "Metropolis"}{ ~~describe this method here }

\item{obj = "SMetropolis"}{ ~~describe this method here }

\item{obj = "PSMetropolis"}{ ~~describe this method here }

\item{obj = "PMetropolis"}{ ~~describe this method here }
}}
\keyword{iteration}
