\name{Umacs-package}
\alias{Umacs-package}
\alias{Umacs}
\docType{package}
\title{Creating Iterative Samplers}
\description{Please refer to the vignette: \code{vignette("Umacs")}.}
\details{\tabular{ll}{Package: \tab Umacs\crVersion: \tab 0.911\crDate: \tab 2007/01/14\crNamespace: \tab Umacs\crEnhances: \tab R2WinBUGS\crDepends: \tab R(>= 1.9.0), methods, utils\crLicense: \tab GPL version 2 or newer\cr}

}
\author{Jouni Kerman <jouni@kerman.com>}
\references{Kerman, Jouni. Umacs: A Universal Markov Chain Sampler. Technical report, Columbia University, New York.}
\keyword{iteration}
